Revert "context TS migration"
